WIP
The current test fails because the carpenter is creating the map (for
some reason) as a linked hash map and not a normal map that we can't
serialise. Added a test to confirm that it's not the carpenter here.

Two questions really, why are we creating it as a linked has map, AFAIK
the carpenter should just be whipping up what it's told to and it's not
being given a linked map

Secondly, should we support serialisation of these objects